pyproject.toml
[project]
name = "liga-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
